refactor: use shared pointer sensor config for all dnd
